Hand fixes for aep and catchments (#1135)
When attempting to create HAND FIM polygons for AEP/Catchments using the
latest HAND FIM version (4.6.1.4), we kept finding that no polygons were
being created. Turns out that line 284 of the hand_fim_processing
lambda_function.py was dropping all of the polygons due to the recently
added `flood_area_above_expected_coeff` column being `nan` for all rows
in the case of static/reference services. This fixes that by dropping
that column completely prior to the `join().dropna()`, but also swaps
out the `join().dropna()` for an inner join, since that seems to be what
was intended there.