New BITS, Bridge Inundation Threat Service, layer inside QC FIM / FIM Pact service layers (#1129)
This is a new set of layers to be added to all services that have a QC
FIM / FIM pact layer. Generally, for all layers that have building data,
this layer will belong in the same services.

The service will be point based with the colors of gray, orange or red
and is dynamically calculated based on the current flow data supporting
that QC FIM service layer.

The data supplied by the FIM dev teams for the bridge points have to key
columns, `threshold_discharge_cfs` and , `threshold_discharge_75_cfs`.
These represent a flow value which can help identify if the bridge is
under no / low risk of flooding, moderate risk of flooding or likely is
inundated. Note: is it easiest to think of the threshold_discharge_cfs
as 100, max discharge calculated by FIM and threshold_discharge_75_cfs
as 75% of the max discharge before being threatened.

The logic scheme for the system is based on using actual flow data
provide dynamically at run time and it's relationship to the two
threshold discharge values above.
- If the actual flow value is above `threshold_discharge_cfs` (100%),
then the point color is red and the message says "High Risk of Flooding"
in the metadata.
- If the actual flow value is below `threshold_discharge_cfs` (100%) and
above `threshold_discharge_75_cfs`, then the point color is orange and
the meta data says "Potentially Flooded"
- If the actual flow is below `threshold_discharge_75_cfs`, then the
point color is gray and meta data says "Low Risk of Flooding".


**Services it is plugged into are:**
- ana_inundation_extent_noaa.mapx
- ana_inundation_extent_hi_noaa.mapx
- ana_inundation_extent_prvi_noaa.mapx
- ana_past_14day_max_inundation_extent_noaa.mapx (layers for both 7 day
and 14 day of course)
- mrf_nbm_10day_max_inundation_extent_noaa.mapx (layers for 3 day, 5 day
and 10 day)
- mrf_gfs_10day_max_inundation_extent_noaa.mapx (layers for 3 day, 5 day
and 10 day)
- rfc_based_5day_max_inundation_extent_noaa.mapx
- srf_18hr_max_inundation_extent_noaa.mapx
- srf_48hr_max_inundation_extent_hi_noaa.mapx
- srf_48hr_max_inundation_extent_prvi_noaa.mapx