Update and renamed ras2fim boundaries service to hecras boundaries service (#1141)
Upgraded and renamed the original `ras2fim boundary` service to `HECRAS
boundary` service which reflects hecras data from both ras2fim and
ripple.

Note: it has a column for 'is_active' which is supposed to be for huc8
records where we are actually using ras2fim/ripple data in our service.
However, while we know all of the huc8s with ras2fim data is included,
we do not know which ripple HUCs were actually used in the service. As a
result, for now, all HUC8s are "is_active" = True. Once the new Ripple
datasets are used, this table will need to be updated.